break up component
